cookbooks/manus-clone/requirements.txt
# swarmkit  # Use local version for testing
-e /Users/ildebrandomagnani/MANUS-API/ai-terminal/packages/sdk-py
python-dotenv
